requirements.txt
##requirement.txt
python-telegram-bot==13.7
configparser
redis
async-timeout==4.0.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==7.0.0
docker==7.0.0
docopt==0.6.2
idna==3.7
packaging==24.0
pipreqs==0.4.13
python-telegram-bot==13.7
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
style==1.1.0
tornado==6.4
tzlocal==5.2
update==0.0.1
urllib3==2.2.1
yarg==0.1.9
APScheduler==3.6.3
async-timeout==4.0.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==7.0.0
docker==7.0.0
docopt==0.6.2
idna==3.7
packaging==24.0
pipreqs==0.4.13
python-telegram-bot==13.7
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
style==1.1.0
tornado==6.4
tzlocal==5.2
update==0.0.1
urllib3==2.2.1
yarg==0.1.9
APScheduler==3.6.3
async-timeout==4.0.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==7.0.0
docker==7.0.0
docopt==0.6.2
idna==3.7
packaging==24.0
pipreqs==0.4.13
python-telegram-bot==13.7
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
style==1.1.0
tornado==6.4
tzlocal==5.2
update==0.0.1
urllib3==2.2.1
yarg==0.1.9
